[modify] floatを使用していた部分をdoubleに置き換え

+++ include/Camera.h
@@ -10,10 +10,10 @@ class Camera
 	Vec3  x_axis_;
 	Vec3  y_axis_;
 	Vec3  z_axis_;
-	float near_clip_;
+	double near_clip_;
 
 public:
-	Camera(Vec3 position, Vec3 direction, Vec3 up, float aspect_ratio, float near_clip)
+	Camera(Vec3 position, Vec3 direction, Vec3 up, double aspect_ratio, double near_clip)
 		:position_(position), near_clip_(near_clip)
 	{
 		//スクリーンの縦横方向の基底ベクトル
@@ -22,7 +22,7 @@ public:
 		z_axis_    = direction;
 	}
 
-	Ray emit(float u, float v)
+	Ray emit(double u, double v)
 	{
 		Vec3 ray_direction = x_axis_ * u + y_axis_ * v + z_axis_;
 		Vec3 ray_position = position_ + ray_direction * near_clip_;
