+++ include/MathLib.h
@@ -2,11 +2,11 @@
 
 namespace LtRenderer
 {
-	inline float pow2(const float value) { return value * value; }
-	inline float pow3(const float value) { return value * value * value; }
-	inline float pow4(const float value) { return value * value * value * value; }
-	inline float pow5(const float value) { return value * value * value * value * value; }
-	inline float saturate(const float value) { return value > 1.0 ? 1.0 : value < 0.0 ? 0.0 : value; }
-	inline float min(const float value1, float value2) { return value1 > value2 ? value2 : value1; }
-	inline float max(const float value1, float value2) { return value1 > value2 ? value1 : value2; }
+	inline double pow2(const double value) { return value * value; }
+	inline double pow3(const double value) { return value * value * value; }
+	inline double pow4(const double value) { return value * value * value * value; }
+	inline double pow5(const double value) { return value * value * value * value * value; }
+	inline double saturate(const double value) { return value > 1.0 ? 1.0 : value < 0.0 ? 0.0 : value; }
+	inline double min(const double value1, double value2) { return value1 > value2 ? value2 : value1; }
+	inline double max(const double value1, double value2) { return value1 > value2 ? value1 : value2; }
 }
